requirements.txt
hyperliquid-python-sdk
requests
pandas
